Separated the rpc methods from the index.js
